test(view): 예외 처리 테스트와 입력 테스트 구분
